third_party/blink/web_tests/fast/xsl/xslt/xslt-string-parameters-quirk-expected.txt

CONSOLE ERROR: Uncaught ReferenceError: XSLTProcessor is not defined
This tests that passing string parameters to the XSLTProcessor works as expected. If this test is successful, the text "SUCCESS" will be shown below.

FAILURE
